feat(wip): move combination analysis logic to vue component
